requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
textblob>=0.17.1
django-cors-headers>=4.0.0
python-decouple>=3.8
gunicorn>=21.2.0
whitenoise>=6.5.0

